--- MyProject/ClientSample/Assets/Script/Network/Datas.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b5067fb5808a48d5b3d9e27b0f76144
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
